fix: avoid using GitHub Personal Access Token while publishing a new release (#30)
Why: force open bump version PRs as GitHub bot

Why we were using GH PACs before: it is the recommendation by Release please to be able to run further GitHub Actions triggered by the PR opened by Release please. However, we are not interested in that behavior because we do not have CI test workflows in this repository, and even in that case, that tests would have already run in their corresponding PRs.

PS: Sorry for using `fix` SemVer type. It should be a `ci`, but we are testing out the automated release process and need to trigger it to create a PR 🙏